+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,0 +1,28 @@
+import {Component, DoCheck, OnInit} from '@angular/core';
+import {CartService} from '../../services/cart.service';
+import {ProductModel} from '../../../products/models/product.model';
+
+@Component({
+  selector: 'app-cart-list',
+  templateUrl: './cart-list.component.html',
+  styleUrls: ['./cart-list.component.less']
+})
+export class CartListComponent implements OnInit, DoCheck {
+
+  products: ProductModel[];
+  totalPrice: number;
+
+  constructor(public cartService: CartService) { }
+
+  ngOnInit() {
+
+  }
+
+  ngDoCheck() {
+    this.products = this.cartService.products;
+    this.totalPrice = this.cartService.totalPrice();
+
+  }
+
+
+}