+++ src/app/products/components/product/product.component.ts
@@ -1,0 +1,26 @@
+import {ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {ProductModel} from '../../models/product.model';
+
+@Component({
+  selector: 'app-product',
+  templateUrl: './product.component.html',
+  styleUrls: ['./product.component.less'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class ProductComponent implements OnInit {
+  @Input()
+  product: ProductModel;
+
+  @Output()
+  addToCart: EventEmitter<void> = new EventEmitter<void>();
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+  onBuy(): void {
+    console.log(`I want to buy ${this.product.name}`);
+    this.addToCart.emit();
+  }
+}