+++ src/app/products/models/product.model.ts
@@ -1,6 +1,8 @@
 export class ProductModel {
   constructor(
     public id: number,
-    public name: string
+    public name: string,
+    public isAvailable: boolean,
+    public price: number
   ) { }
 }