+++ src/app/products/services/products.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
-import {ProductModel} from './models/product.model';
+import {ProductModel} from '../models/product.model';
 
 const productList: Array<ProductModel> = [
-  new ProductModel(1, 'Book'),
-  new ProductModel(2, 'Excel'),
-  new ProductModel(3, 'Samsung')
+  new ProductModel(1, 'Book', true, 10),
+  new ProductModel(2, 'Excel', true, 20),
+  new ProductModel(3, 'Samsung', false, 30)
 ];
 
 @Injectable({