+++ changelog.md
@@ -1,3 +1,15 @@
-# Create clear functionality
+# Create functionality
 
+### Home Task 2
+
+- Add OnPush strategy to components;
+- Add possibility to change count of product in CartComponent;
+- Add calculating of totalPrice and updateQuantity to CartService;
+- Add method of addToCart and removeProduct to CartService;
+- Add title text using ViewChild;
+- Add directive [hightlight] for changing background;
+- Buy button is disabled if product is not available;
+- Use pipe currency for price; 
+
+### Home Task 1
 Added ability to clear carts